+++ ModuleUnserializer/Entities/Skill.cs
@@ -10,7 +10,7 @@ namespace ModuleUnserializer.Entities
 	{
 		public string Index;
 		public string Name;
-		public int Flags;
+		public long Flags;
 		public int MaxLevel;
 		public string Description;
 
@@ -24,7 +24,7 @@ namespace ModuleUnserializer.Entities
 			tab.Index = s[j++];
 			tab.Index = tab.Index.Substring(tab.Index.IndexOf("_") + 1);
 			tab.Name = s[j++];
-			tab.Flags = Convert.ToInt32(s[j++]);
+			tab.Flags = Convert.ToInt64(s[j++]);
 			tab.MaxLevel = Convert.ToInt32(s[j++]);
 			tab.Description = s[j++];
 			return tab;
